app_utils.py
import os
from pathlib import Path
from typing import Any

app_dir = Path(__file__).parent
env_file = app_dir / ".env"


def load_dotenv(dotenv_path: os.PathLike[str] = env_file, **kwargs: Any) -> None:
    """
    A convenience wrapper around `dotenv.load_dotenv` that warns if `dotenv` is not installed.
    It also returns `None` to make it easier to ignore the return value.
    """
    try:
        import dotenv

        dotenv.load_dotenv(dotenv_path=dotenv_path, **kwargs)
    except ImportError:
        import warnings

        warnings.warn(
            "Could not import `dotenv`. If you want to use `.env` files to "
            "load environment variables, please install it using "
            "`pip install python-dotenv`.",
            stacklevel=2,
        )


def validate_data(data: Any) -> bool:
    """
    Validate the input data to ensure it meets the required format and quality standards.
    """
    # Implement data validation logic here
    return True


def transform_data(data: Any) -> Any:
    """
    Apply necessary transformations to the data, such as normalization and outlier detection.
    """
    # Implement data transformation logic here
    return data


def preprocess_data(data: Any) -> Any:
    """
    Preprocess the data by handling missing values, data transformation, and outlier detection.
    """
    if not validate_data(data):
        raise ValueError("Invalid data format or quality.")
    
    data = transform_data(data)
    return data
